+++ packages/shed/tests/test_cli_tools_prompt_examples.py
@@ -395,10 +395,10 @@ def test_pipe_passes_previous_stdout_to_next_stdin(
             RunFileCommands(
                 chain="pipe",
                 file_commands=[
-                    RunFileCommand(command="grep", argv=["-n", "-R", "TODO", "src/"])
+                    RunFileCommand(command="grep", argv=["-n", "-r", "TODO", "src/"])
                 ],
             ),
-            [["grep", "-n", "-R", "TODO", "{BASE}/src"]],
+            [["grep", "-n", "-r", "TODO", "{BASE}/src"]],
         ),
         (
             RunFileCommands(
@@ -474,7 +474,7 @@ def test_prompt_examples_real_subprocess_output_semantics(
             RunFileCommands(
                 chain="pipe",
                 file_commands=[
-                    RunFileCommand(command="grep", argv=["-n", "-R", "TODO", "src/"])
+                    RunFileCommand(command="grep", argv=["-n", "-r", "TODO", "src/"])
                 ],
             ),
             "TODO",